Add a new high-level BlackboardAPIClient
Later more code will be added to BlackboardAPIClient and
BasicBlackboardAPIClient will become a private helper. For now this PR
just gets the new high-level BlackboardAPIClient in place and gets a
couple of simple views using it.
