+++ lms/services/__init__.py
@@ -1,5 +1,6 @@
 from lms.services.canvas import CanvasService
 from lms.services.exceptions import (
+    BlackboardFileNotFoundInCourse,
     CanvasAPIError,
     CanvasAPIPermissionError,
     CanvasAPIServerError,
@@ -22,6 +23,10 @@ def includeme(config):
     config.register_service_factory("lms.services.http.factory", name="http")
     config.register_service_factory(
         "lms.services.basic_blackboard_api.factory", name="basic_blackboard_api_client"
+    )
+    config.register_service_factory(
+        "lms.services.blackboard_api.blackboard_api_client_factory",
+        name="blackboard_api_client",
     )
     config.register_service_factory(
         "lms.services.canvas_api.canvas_api_client_factory", name="canvas_api_client"
