+++ tests/unit/lms/services/blackboard_api/_schemas_test.py
@@ -1,10 +1,10 @@
 import pytest
 
-from lms.validation._exceptions import ValidationError
-from lms.views.api.blackboard._schemas import (
+from lms.services.blackboard_api._schemas import (
     BlackboardListFilesSchema,
     BlackboardPublicURLSchema,
 )
+from lms.validation._exceptions import ValidationError
 from tests import factories
