+++ tests/unit/services.py
@@ -7,6 +7,7 @@ from lms.services import CanvasService
 from lms.services.application_instance import ApplicationInstanceService
 from lms.services.assignment import AssignmentService
 from lms.services.basic_blackboard_api import BasicBlackboardAPIClient
+from lms.services.blackboard_api.client import BlackboardAPIClient
 from lms.services.canvas_api import CanvasAPIClient
 from lms.services.course import CourseService
 from lms.services.file import FileService
@@ -31,6 +32,7 @@ __all__ = (
     "application_instance_service",
     "assignment_service",
     "basic_blackboard_api_client",
+    "blackboard_api_client",
     "canvas_api_client",
     "canvas_service",
     "course_service",
@@ -99,6 +101,11 @@ def basic_blackboard_api_client(mock_service):
     return mock_service(
         BasicBlackboardAPIClient, service_name="basic_blackboard_api_client"
     )
+
+
+@pytest.fixture
+def blackboard_api_client(mock_service):
+    return mock_service(BlackboardAPIClient, service_name="blackboard_api_client")
 
 
 @pytest.fixture
